requirements.txt
loguru~=0.6.0
fastapi~=0.85.2
config~=0.5.1
starlette~=0.20.4
uvicorn~=0.19.0
python-dotenv~=0.21.0